backend/requirements.txt
asgiref==3.9.1
Django==4.2.23
djangorestframework==3.14.0
sqlparse==0.5.3
style==1.1.0
typing_extensions==4.15.0
update==0.0.1
python-jose[cryptography]==3.3.0
httpx==0.27.0
python-dotenv==1.0.1
dj-database-url==2.2.0
django-cors-headers==4.4.0

# Composio SDK (Python)
composio
composio_langchain

# Streaming / utilities
orjson==3.10.7
pydantic==2.9.2
urllib3<2

# Database drivers
psycopg[binary]>=3.1

# LangChain provider (to be used in chat runtime later)
langchain
langchain-openai
openai
langgraph
langchain-community
pypdf
python-docx

# Realtime + background workers
channels==4.1.0
channels-redis==4.2.0
redis==5.0.7
celery==5.4.0
uvicorn[standard]==0.30.6
whitenoise==6.7.0
